+++ src/lib/menu.c
@@ -62,7 +62,7 @@ dup_menu_item(ncmenu_int_item* dst, const struct ncmenu_item* src){
   }
   sdup[n + mbbytes] = '\0';
   dst->shortdesc = sdup;
-  dst->shortdesccols = mbswidth(dst->shortdesc);
+  dst->shortdesccols = ncstrwidth(dst->shortdesc);
   return 0;
 #undef CTLMOD
 #undef ALTMOD
@@ -94,7 +94,7 @@ dup_menu_section(ncmenu_int_section* dst, const struct ncmenu_section* src){
         return -1;
       }
       gotitem = true;
-      int cols = mbswidth(dst->items[i].desc);
+      int cols = ncstrwidth(dst->items[i].desc);
       if(dst->items[i].shortdesc){
         cols += 2 + dst->items[i].shortdesccols; // two spaces minimum
       }
@@ -140,7 +140,7 @@ dup_menu_sections(ncmenu* ncm, const ncmenu_options* opts, int* totalwidth, int*
   int i;
   for(i = 0 ; i < opts->sectioncount ; ++i){
     if(opts->sections[i].name){
-      int cols = mbswidth(opts->sections[i].name);
+      int cols = ncstrwidth(opts->sections[i].name);
       if(rightaligned){ // FIXME handle more than one right-aligned section
         ncm->sections[i].xoff = -(cols + 2);
       }else{
@@ -212,14 +212,14 @@ section_x(const ncmenu* ncm, int x){
       if(x < pos){
         break;
       }
-      if(x < pos + mbswidth(ncm->sections[i].name)){
+      if(x < pos + ncstrwidth(ncm->sections[i].name)){
         return i;
       }
     }else{
       if(x < ncm->sections[i].xoff){
         break;
       }
-      if(x < ncm->sections[i].xoff + mbswidth(ncm->sections[i].name)){
+      if(x < ncm->sections[i].xoff + ncstrwidth(ncm->sections[i].name)){
         return i;
       }
     }
@@ -274,7 +274,7 @@ write_header(ncmenu* ncm){ ncm->ncp->channels = ncm->headerchannels;
         }
         cell_release(ncm->ncp, &cl);
       }
-      xoff += mbswidth(ncm->sections[i].name);
+      xoff += ncstrwidth(ncm->sections[i].name);
     }
   }
   while(xoff < dimx){