+++ src/lib/reader.c
@@ -178,7 +178,7 @@ int ncreader_move_down(ncreader* n){
 
 // only writing can enlarge the textarea. movement can pan, but not enlarge.
 int ncreader_write_egc(ncreader* n, const char* egc){
-  const int cols = mbswidth(egc);
+  const int cols = ncstrwidth(egc);
   if(cols < 0){
     logerror(n->ncp->nc, "Fed illegal UTF-8 [%s]\n", egc);
     return -1;